fix(services): use `.merge` to add `/healthz` route in axum
The `static-files` example was not working correctly in the
previous commit. The `static-files` example  was using `SpaRouter` from
`axum_extra::routing::SpaRouter` which responds to all unmatched
routes with the `index.html` page. To fix this i used the `.merge` method
in axum, instead of simply adding `/healthz` route with a
`.route` method.