Add --allow-dirty to command hint at end of init (#1076)
This commit modifies the command hint `cargo cyndra deploy` that
appears at the end of `cargo cyndra init` so that it includes the
`--allow-dirty` flag.

Without this flag, `cargo cyndra` will fail every time a newbie tries
to init a new project because cyndra adds files to the project but
doesn't commit them to git. cyndra then notices that there are
uncommitted files, and doesn't allow the newbie to deploy the project
without specifically allowing the dirty files. This is a bit of a bad
experience, because the newbie is just following instructions but also
being told that those instructions were wrong.